+++ FlatFiles/IWriter.cs
@@ -37,6 +37,22 @@ namespace FlatFiles
         /// <param name="values">The values to write.</param>
         /// <returns>The textual representation of the given values.</returns>
         Task WriteAsync(object[] values);
+
+        /// <summary>
+        /// Write the given data directly to the output. By default, this will
+        /// not include a newline.
+        /// </summary>
+        /// <param name="data">The data to write to the output.</param>
+        /// <param name="writeRecordSeparator">Indicates whether a record separator should be written after the data.</param>
+        void WriteRaw(String data, bool writeRecordSeparator = false);
+
+        /// <summary>
+        /// Write the given data directly to the output. By default, this will
+        /// not include a newline.
+        /// </summary>
+        /// <param name="data">The data to write to the output.</param>
+        /// <param name="writeRecordSeparator">Indicates whether a record separator should be written after the data.</param>
+        Task WriteRawAsync(String data, bool writeRecordSeparator = false);
     }
 
     internal interface IWriterWithMetadata : IWriter