+++ FlatFiles/SeparatedValueRecordWriter.cs
@@ -194,5 +194,15 @@ namespace FlatFiles
         {
             await writer.WriteAsync(Metadata.ExecutionContext.Options.RecordSeparator ?? Environment.NewLine).ConfigureAwait(false);
         }
+
+        public void WriteRaw(String data)
+        {
+            writer.Write(data);
+        }
+
+        public Task WriteRawAsync(String data)
+        {
+            return writer.WriteAsync(data);
+        }
     }
 }